+++ internal/dbadaptertest/test_adapter.go
@@ -19,6 +19,7 @@ type TestAdapter struct {
 	mu                 sync.Mutex
 
 	JobCancelCalled              bool
+	JobCancelTxCalled            bool
 	JobInsertCalled              bool
 	JobInsertTxCalled            bool
 	JobInsertManyCalled          bool
@@ -30,6 +31,7 @@ type TestAdapter struct {
 	LeadershipResignedCalled     bool
 
 	JobCancelFunc              func(ctx context.Context, id int64) (*dbadapter.JobCancelResult, error)
+	JobCancelTxFunc            func(ctx context.Context, tx pgx.Tx, id int64) (*dbadapter.JobCancelResult, error)
 	JobInsertFunc              func(ctx context.Context, params *dbadapter.JobInsertParams) (*dbadapter.JobInsertResult, error)
 	JobInsertTxFunc            func(ctx context.Context, tx pgx.Tx, params *dbadapter.JobInsertParams) (*dbadapter.JobInsertResult, error)
 	JobInsertManyFunc          func(ctx context.Context, params []*dbadapter.JobInsertParams) (int64, error)
@@ -46,6 +48,16 @@ func (ta *TestAdapter) JobCancel(ctx context.Context, id int64) (*dbadapter.JobC
 
 	if ta.JobCancelFunc != nil {
 		return ta.JobCancelFunc(ctx, id)
+	}
+
+	return ta.fallthroughAdapter.JobCancel(ctx, id)
+}
+
+func (ta *TestAdapter) JobCancelTx(ctx context.Context, tx pgx.Tx, id int64) (*dbadapter.JobCancelResult, error) {
+	ta.atomicSetBoolTrue(&ta.JobCancelTxCalled)
+
+	if ta.JobCancelTxFunc != nil {
+		return ta.JobCancelTxFunc(ctx, tx, id)
 	}
 
 	return ta.fallthroughAdapter.JobCancel(ctx, id)